Correct background coloring bugs in Hackathon page
When loading, because some stylesheets did not specify the background color,
it would inherit the blue color from the SEI pages and display it while the background image was loading.
A similar event  happened in the header stylesheet because the background was set to a darker color than the
background image's background.