Correct background coloring bugs in Hackathon page (#67)
When loading, because some stylesheets did not specify the background colour, it would inherit the
blue colour from the SEI pages and display it while the background image was loading. A similar event happened in the header stylesheet because the background was set to a darker colour than the background image's background.